Extract permissions related functions from LibraryManager
