+++ plugins/org.eclipse.fordiac.ide.library/src/org/eclipse/fordiac/ide/library/LibraryManager.java
@@ -19,7 +19,6 @@ import java.io.OutputStream;
 import java.nio.file.DirectoryStream.Filter;
 import java.nio.file.FileSystems;
 import java.nio.file.FileVisitResult;
-import java.nio.file.FileVisitor;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -28,9 +27,6 @@ import java.nio.file.StandardWatchEventKinds;
 import java.nio.file.WatchKey;
 import java.nio.file.WatchService;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.nio.file.attribute.DosFileAttributeView;
-import java.nio.file.attribute.PosixFileAttributeView;
-import java.nio.file.attribute.PosixFilePermission;
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -141,7 +137,7 @@ public enum LibraryManager {
 			FordiacLogHelper.logError("Cannot register watch watch service!", e); //$NON-NLS-1$
 		}
 
-		setStandardLibsReadOnly();
+		LibraryPermission.setLibReadOnly(standardLibraryPath);
 	}
 
 	/**
@@ -179,9 +175,9 @@ public enum LibraryManager {
 	}
 
 	/**
-	 * Initialise map with all libraries contained in the folder specified
+	 * Initialize map with all libraries contained in the folder specified
 	 *
-	 * @param map      map to initialise
+	 * @param map      map to initialize
 	 * @param path     path to folder
 	 * @param standard if libraries are standard
 	 * @param baseURI  URI to use as base
@@ -298,7 +294,7 @@ public enum LibraryManager {
 							fileOutputStream.write(buffer, 0, len);
 						}
 					}
-					setPathReadOnly(newFile);
+					LibraryPermission.setPathReadOnly(newFile);
 				}
 				entry = zipInputStream.getNextEntry();
 			}
@@ -324,14 +320,14 @@ public enum LibraryManager {
 			Files.walkFileTree(folder, new SimpleFileVisitor<Path>() {
 				@Override
 				public FileVisitResult visitFile(final Path file, final BasicFileAttributes attrs) throws IOException {
-					setPathEditable(file);
+					LibraryPermission.setPathEditable(file);
 					Files.delete(file);
 					return FileVisitResult.CONTINUE;
 				}
 
 				@Override
 				public FileVisitResult postVisitDirectory(final Path dir, final IOException exc) throws IOException {
-					setPathEditable(dir);
+					LibraryPermission.setPathEditable(dir);
 					Files.delete(dir);
 					return FileVisitResult.CONTINUE;
 				}
@@ -356,112 +352,6 @@ public enum LibraryManager {
 			throw new IOException("Entry is outside of the target dir: " + zipEntry.getName()); //$NON-NLS-1$
 		}
 		return destPath;
-	}
-
-	/**
-	 * Sets a specific path read only
-	 *
-	 * @param path Path to set read only
-	 */
-	private static void setPathReadOnly(final Path path) {
-		final DosFileAttributeView dosView = Files.getFileAttributeView(path, DosFileAttributeView.class);
-		if (dosView != null) {
-			try {
-				dosView.setReadOnly(true);
-			} catch (final IOException e) {
-				// empty
-			}
-		}
-		final PosixFileAttributeView posixView = Files.getFileAttributeView(path, PosixFileAttributeView.class);
-		if (posixView != null) {
-			try {
-				final Set<PosixFilePermission> permissions = posixView.readAttributes().permissions();
-				permissions.remove(PosixFilePermission.OWNER_WRITE);
-				permissions.remove(PosixFilePermission.GROUP_WRITE);
-				permissions.remove(PosixFilePermission.OTHERS_WRITE);
-				posixView.setPermissions(permissions);
-			} catch (final IOException e) {
-				// empty
-			}
-		}
-	}
-
-	/**
-	 * Sets a specific path editable
-	 *
-	 * @param path Path to set editable
-	 */
-	private static void setPathEditable(final Path path) {
-		final DosFileAttributeView dosView = Files.getFileAttributeView(path, DosFileAttributeView.class);
-		if (dosView != null) {
-			try {
-				dosView.setReadOnly(false);
-			} catch (final IOException e) {
-				// empty
-			}
-		}
-		final PosixFileAttributeView posixView = Files.getFileAttributeView(path, PosixFileAttributeView.class);
-		if (posixView != null) {
-			try {
-				final Set<PosixFilePermission> permissions = posixView.readAttributes().permissions();
-				permissions.add(PosixFilePermission.OWNER_WRITE);
-				permissions.add(PosixFilePermission.GROUP_WRITE);
-				permissions.add(PosixFilePermission.OTHERS_WRITE);
-				posixView.setPermissions(permissions);
-			} catch (final IOException e) {
-				// empty
-			}
-		}
-	}
-
-	private void setStandardLibsReadOnly() {
-		final WorkspaceJob job = new WorkspaceJob(Messages.LibraryManager_SetStandardLibrariesReadOnly) {
-
-			@Override
-			public IStatus runInWorkspace(final IProgressMonitor monitor) throws CoreException {
-
-				try {
-					Files.walkFileTree(standardLibraryPath, new FileVisitor<Path>() {
-
-						@Override
-						public FileVisitResult preVisitDirectory(final Path dir, final BasicFileAttributes attrs)
-								throws IOException {
-							return FileVisitResult.CONTINUE;
-						}
-
-						@Override
-						public FileVisitResult visitFile(final Path file, final BasicFileAttributes attrs)
-								throws IOException {
-							setPathReadOnly(file);
-							return FileVisitResult.CONTINUE;
-						}
-
-						@Override
-						public FileVisitResult visitFileFailed(final Path file, final IOException exc)
-								throws IOException {
-							return FileVisitResult.CONTINUE;
-						}
-
-						@Override
-						public FileVisitResult postVisitDirectory(final Path dir, final IOException exc)
-								throws IOException {
-							return FileVisitResult.CONTINUE;
-						}
-					});
-				} catch (final IOException e) {
-					// empty
-				}
-				return Status.OK_STATUS;
-			}
-
-			@Override
-			public boolean belongsTo(final Object family) {
-				return family == FAMILY_FORDIAC_LIBRARY;
-			}
-		};
-		job.setRule(null);
-		job.setPriority(Job.DECORATE);
-		job.schedule();
 	}
 
 	/**
